Refactor factory metods and adjusts error handling, extract randomization logic into pkg randstr
